requirements.txt
# Transit Service Reliability & Demand Planning System - Python Dependencies

# AWS SDK
boto3>=1.28.0
botocore>=1.31.0

# HTTP requests
requests>=2.31.0

# Snowflake
snowflake-connector-python>=3.5.0
snowflake-snowpark-python>=1.8.0

# dbt
dbt-core>=1.5.0
dbt-snowflake>=1.5.0

# ML/Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# OpenAI (for chatbot)
openai>=1.0.0

# Flask (for chatbot API)
flask>=2.3.0

# Configuration
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

